+++ interfaces/global-state-types.ts
@@ -1,9 +1,13 @@
 import { User } from "@supabase/supabase-js";
 
 export interface GlobalStateInterface {
-  user?: User | null;
+  user: User | null;
   onboarded?: boolean;
   waitlisted?: boolean;
   range?: number;
   insightRepoLimit?: number;
+  sessionToken?: string | null;
+  providerToken?: string | null;
+  userId?: number | null;
+  hasReports?: boolean;
 }