+++ tailwind.config.js
@@ -23,7 +23,8 @@ module.exports = {
       },
       colors: {
         sauced: {
-          orange: "hsla(19, 100%, 50%, 1)"
+          orange: "hsla(19, 100%, 50%, 1)",
+          light: "hsl(24, 100%, 95%)"
         },
         gradient: {
           "dark-two": "hsl(26, 54%, 15%)",